heddle#327: re-ground spike PoC against real init.rs types
Codex r1 (PR #361): the PoC was mirror-shaped — it measured a simplified
InitSchema-style struct, so the spike's drift/discriminator numbers and "zero
drift / migration risk ≈ 0" conclusions didn't reflect the production surface.
Re-ground the whole PoC against the REAL init.rs types and make the
measurements assert, then align the doc's evidence:

- Output struct now replicates the real private InitOutput field-for-field,
  INCLUDING the #[serde(skip_serializing)] #[serde(rename="verification")]
  trust field. schemars re-exposes it as a required `writeOnly` property the
  wire bytes never emit — the migration is NOT semantics-free.
- Input struct modelled as the real clap::Args InitArgs wired through a
  Subcommand enum (Parser on Cli, Args on the leaf), not a standalone Parser.
- measure.rs ASSERTS the discriminator gap (!schemars_pinned && custom_pinned),
  the skip_serializing drift (schemars schema − wire output == {verification}),
  and the typed-example vs curated-doc-sample divergence — no longer print-only.
- Split lib.rs (394 lines) into args/output/schemars_path/custom_path modules,
  each well under the 300-line limit.
- Doc: correct the zero-drift and "one comment feeds both surfaces" overclaims;
  require a JsonSchema bound (a derive can't add a sibling derive); keep leaf
  args on clap::Args. Recommendation (schemars for v1) stands — the per-verb
  cost is now stated accurately (#[schemars(skip)] + const helper + doc-sample
  rebaseline), not hand-waved as zero-touch.
