+++ crates/cli-macro-poc/tests/measure.rs
@@ -2,8 +2,12 @@
 //! Measurement harness for the heddle#327 spike. These tests ARE the
 //! "measure output quality" deliverable: they prove both emitters cover the
 //! documented contract, prove the attribute-design claims (doc comments →
-//! descriptions, example carry-through), and print the comparison table the
-//! spike doc cites. Run with `--nocapture` to see the table.
+//! descriptions, example carry-through), and — crucially — ASSERT the two
+//! schemars costs the decision doc cites: the unpinned `output_kind`
+//! discriminator and the phantom `verification` property schemars re-introduces
+//! from the real `#[serde(skip_serializing)]` `trust` field. The numbers in the
+//! doc are guarded by these assertions, not merely printed. Run with
+//! `--nocapture` to also see the comparison table.
 
 use std::collections::BTreeSet;
 
@@ -11,6 +15,23 @@ use heddle_cli_macro_poc::{
     custom_path, documented_sample_keys, init_example, property_keys, schemars_path,
 };
 use serde_json::Value;
+
+/// `output_kind` is "pinned" if the schema constrains it to a single value via
+/// `const` or a one-element `enum`.
+fn output_kind_pinned(schema: &Value) -> bool {
+    schema.pointer("/properties/output_kind/const").is_some()
+        || schema.pointer("/properties/output_kind/enum").is_some()
+}
+
+fn serialized_example_keys() -> BTreeSet<String> {
+    serde_json::to_value(init_example())
+        .unwrap()
+        .as_object()
+        .expect("init_example serializes to an object")
+        .keys()
+        .cloned()
+        .collect()
+}
 
 /// The drift contract: every key the documented sample asserts must appear as a
 /// property in BOTH emitters' schemas. This is the property `heddle doctor
@@ -20,9 +41,8 @@ use serde_json::Value;
 fn both_emitters_cover_the_documented_sample_keys() {
     let documented: BTreeSet<&str> = documented_sample_keys().into_iter().collect();
 
-    let schemars_keys: BTreeSet<String> = property_keys(&schemars_path::schema())
-        .into_iter()
-        .collect();
+    let schemars_keys: BTreeSet<String> =
+        property_keys(&schemars_path::schema()).into_iter().collect();
     let custom_keys: BTreeSet<String> = property_keys(&custom_path::schema()).into_iter().collect();
 
     for key in &documented {
@@ -90,7 +110,120 @@ fn custom_emitter_embeds_the_documented_example() {
     assert_eq!(*example, serde_json::to_value(init_example()).unwrap());
 }
 
-/// Print the comparison table the spike doc cites. Not an assertion — a probe.
+/// Spike answer Q2 — the measured discriminator gap, ASSERTED (not just
+/// printed). schemars renders `output_kind` as a bare `{"type":"string"}` from
+/// a plain `&'static str`; the custom emitter pins it to `"const":"init"`. The
+/// decision doc treats this as the single concrete schemars cost and proposes a
+/// #205 helper for it — so the gate must fail if a schemars upgrade starts
+/// emitting a const/enum here (making the helper unnecessary) or the custom
+/// emitter stops pinning it.
+#[test]
+fn discriminator_gap_is_real_schemars_unpinned_custom_pinned() {
+    let schemars_pinned = output_kind_pinned(&schemars_path::schema());
+    let custom_pinned = output_kind_pinned(&custom_path::schema());
+    assert!(
+        !schemars_pinned,
+        "schemars unexpectedly pinned output_kind — the discriminator gap (and \
+         the proposed #205 helper) is now stale; revisit the decision doc"
+    );
+    assert!(
+        custom_pinned,
+        "custom emitter stopped pinning output_kind as a const"
+    );
+}
+
+/// Spike answer Q2 — the skip_serializing drift, ASSERTED. The real
+/// `InitOutput.trust` field is `#[serde(skip_serializing)] #[serde(rename =
+/// "verification")]`. serde drops it from the wire bytes; schemars' derive
+/// re-introduces a `verification` property (as `writeOnly`) — and even lists it
+/// as REQUIRED. So deriving `JsonSchema` on the real struct is NOT
+/// semantics-free: the schema gains a required property the command never
+/// emits, and `doctor schemas` (keys-in-sample only) would not catch it. The
+/// custom emitter, walking the serialized field set, omits it by construction.
+#[test]
+fn schemars_re_exposes_skip_serialized_verification_field() {
+    let schemars_schema = schemars_path::schema();
+    let schemars_props: BTreeSet<String> =
+        property_keys(&schemars_schema).into_iter().collect();
+    let custom_props: BTreeSet<String> =
+        property_keys(&custom_path::schema()).into_iter().collect();
+    let wire_keys = serialized_example_keys();
+
+    assert!(
+        schemars_props.contains("verification"),
+        "schemars should re-expose the skip_serializing `verification` field; \
+         got props: {schemars_props:?}"
+    );
+    let required: BTreeSet<String> = schemars_schema
+        .get("required")
+        .and_then(Value::as_array)
+        .map(|a| a.iter().filter_map(|v| v.as_str().map(str::to_owned)).collect())
+        .unwrap_or_default();
+    assert!(
+        required.contains("verification"),
+        "schemars marks the phantom `verification` field REQUIRED — the drift is \
+         worse than optional; got required: {required:?}"
+    );
+    assert!(
+        !wire_keys.contains("verification"),
+        "the serialized init output must NOT carry `verification` (it is \
+         skip_serializing); got wire keys: {wire_keys:?}"
+    );
+    assert!(
+        !custom_props.contains("verification"),
+        "the custom emitter walks the serialized field set and must omit \
+         `verification`; got props: {custom_props:?}"
+    );
+    // The drift is exactly this one phantom property.
+    let schemars_minus_wire: BTreeSet<&String> = schemars_props.difference(&wire_keys).collect();
+    assert_eq!(
+        schemars_minus_wire,
+        BTreeSet::from([&"verification".to_string()]),
+        "schemars schema vs wire output should differ by exactly `verification`"
+    );
+}
+
+/// Spike answer Q1/examples — the typed example diverges from the curated
+/// `docs/json-schemas.md` sample, ASSERTED. The real `InitOutput` always
+/// serializes the principal fields (no `skip_serializing_if`), so a faithful
+/// typed example carries keys the hand-curated prose sample omits. This is the
+/// evidence for "typed examples beat prose samples": the example cannot drift
+/// from the struct, but the prose sample already has. heddle#205 must rebaseline
+/// the documented sample to the real shape.
+#[test]
+fn typed_example_diverges_from_curated_doc_sample() {
+    let documented: BTreeSet<String> = documented_sample_keys()
+        .into_iter()
+        .map(str::to_owned)
+        .collect();
+    let wire_keys = serialized_example_keys();
+
+    // The real output is a superset of the documented sample's keys.
+    assert!(
+        documented.is_subset(&wire_keys),
+        "every documented key must be present in the real output; documented: \
+         {documented:?}, wire: {wire_keys:?}"
+    );
+
+    // …and carries exactly these extra always-serialized fields the curated
+    // sample drops. Pinning the set turns Codex's "example diverges from docs"
+    // finding into a measured regression guard.
+    let extra: BTreeSet<String> = wire_keys.difference(&documented).cloned().collect();
+    let expected_extra = BTreeSet::from([
+        "principal".to_string(),
+        "principal_recommended_action".to_string(),
+        "principal_source".to_string(),
+        "principal_status".to_string(),
+    ]);
+    assert_eq!(
+        extra, expected_extra,
+        "typed example vs curated doc sample should differ by exactly the \
+         always-serialized principal fields"
+    );
+}
+
+/// Print the comparison table the spike doc cites. Not an assertion — a probe;
+/// the gap and drift it shows are asserted by the dedicated tests above.
 #[test]
 fn print_measurement_table() {
     let s = schemars_path::schema();
@@ -103,38 +236,38 @@ fn print_measurement_table() {
     let c_keys = property_keys(&c).len();
 
     let s_has_defs = s.get("definitions").is_some() || s.get("$defs").is_some();
-    let s_const = s
-        .pointer("/properties/output_kind/const")
-        .or_else(|| s.pointer("/properties/output_kind/enum"))
-        .is_some();
-    let c_const = c.pointer("/properties/output_kind/const").is_some();
+    let s_const = output_kind_pinned(&s);
+    let c_const = output_kind_pinned(&c);
+    let s_props: BTreeSet<String> = property_keys(&s).into_iter().collect();
+    let s_has_verification = s_props.contains("verification");
 
     println!("\n=== heddle#327 PoC measurement — `init` verb, both ways ===");
-    println!("{:<34} | {:>10} | {:>10}", "metric", "schemars", "custom");
-    println!("{:-<34}-+-{:-<10}-+-{:-<10}", "", "", "");
+    println!("{:<36} | {:>10} | {:>10}", "metric", "schemars", "custom");
+    println!("{:-<36}-+-{:-<10}-+-{:-<10}", "", "", "");
     println!(
-        "{:<34} | {:>10} | {:>10}",
+        "{:<36} | {:>10} | {:>10}",
         "pretty-printed schema bytes",
         s_pretty.len(),
         c_pretty.len()
     );
     println!(
-        "{:<34} | {:>10} | {:>10}",
+        "{:<36} | {:>10} | {:>10}",
         "top-level property keys", s_keys, c_keys
     );
     println!(
-        "{:<34} | {:>10} | {:>10}",
+        "{:<36} | {:>10} | {:>10}",
         "uses $ref/definitions (nested)", s_has_defs, false
     );
     println!(
-        "{:<34} | {:>10} | {:>10}",
+        "{:<36} | {:>10} | {:>10}",
         "output_kind pinned (const/enum)", s_const, c_const
     );
-    println!("{:<34} | {:>10} | {:>10}", "carries example", true, true);
     println!(
-        "{:<34} | {:>10} | {:>10}",
-        "needs schemars dep", true, false
+        "{:<36} | {:>10} | {:>10}",
+        "phantom `verification` property", s_has_verification, false
     );
+    println!("{:<36} | {:>10} | {:>10}", "carries example", true, true);
+    println!("{:<36} | {:>10} | {:>10}", "needs schemars dep", true, false);
     println!("===========================================================\n");
 
     println!("--- schemars schema ---\n{s_pretty}\n");
